Jaldi

+++ toko_online/application/controllers/dashboard.php
@@ -52,10 +52,16 @@ class Dashboard extends CI_Controller {
 
     public function proses_pesanan()
     {
+        $is_processed = $this->model_invoices->index();
+        if($is_processed){
         $this->cart->destroy();
         $this->load->view('templates/header');
         $this->load->view('templates/sidebar');
         $this->load->view('proses_pesanan');
         $this->load->view('templates/footer');
+        } else{
+                echo "Maaf Pesanan Anda Gagal diproses";
+        }
+        
     }
 }
